making requested changes to PR
